scraper/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
lxml==4.9.3
